Expose Match and Allow `require` in Scripts (0.1.16)
This patch exposes `match` and performs a light refactor to prevent a circular dependency from saddle to its libraries and back. We also now allow `require` including relatives in saddle scripts.
