+++ src/cli/commands/test.ts
@@ -5,7 +5,7 @@ import jest from 'jest';
 import path from 'path';
 import {loadConfig} from '../../config';
 
-import {info, debug, warn, error} from '../logger';
+import {info, debug, warn, error} from '../../logger';
 
 export async function test(argv: yargs.Arguments, coverage: boolean, verbose: number): Promise<void> {
   info(`Saddle: running contract ${coverage ? 'coverage' : 'tests'} with jest...\n`, verbose);
